feat: add another check for nir correctness
