+++ backend/api/db/models/nir.py
@@ -1,0 +1,12 @@
+def nir_format(field: str):
+    if field:
+        nir = field.strip()
+        if len(nir) != 15:
+            return "The NIR must be 15 digit long"
+        if not all([char.isdigit() for char in nir]):
+            return "The NIR cannot contain letters"
+        payload = int(nir[:13])
+        check = int(nir[13:])
+        valid = 97 - payload % 97 == check
+        if not valid:
+            return "The NIR provided has invalid format"
